Split CAPI into different files and modify date/time/timestamp values
