+++ src/main/capi/duckdb-c.cpp
@@ -1,0 +1,108 @@
+#include "duckdb/main/capi_internal.hpp"
+
+using duckdb::DuckDB;
+using duckdb::Connection;
+using duckdb::DatabaseData;
+using duckdb::DBConfig;
+
+duckdb_state duckdb_open_ext(const char *path, duckdb_database *out, duckdb_config config, char **error) {
+	auto wrapper = new DatabaseData();
+	try {
+		auto db_config = (DBConfig *)config;
+		wrapper->database = duckdb::make_unique<DuckDB>(path, db_config);
+	} catch (std::exception &ex) {
+		if (error) {
+			*error = strdup(ex.what());
+		}
+		delete wrapper;
+		return DuckDBError;
+	} catch (...) { // LCOV_EXCL_START
+		if (error) {
+			*error = strdup("Unknown error");
+		}
+		delete wrapper;
+		return DuckDBError;
+	} // LCOV_EXCL_STOP
+	*out = (duckdb_database)wrapper;
+	return DuckDBSuccess;
+}
+
+duckdb_state duckdb_open(const char *path, duckdb_database *out) {
+	return duckdb_open_ext(path, out, nullptr, nullptr);
+}
+
+void duckdb_close(duckdb_database *database) {
+	if (*database) {
+		auto wrapper = (DatabaseData *)*database;
+		delete wrapper;
+		*database = nullptr;
+	}
+}
+
+duckdb_state duckdb_connect(duckdb_database database, duckdb_connection *out) {
+	auto wrapper = (DatabaseData *)database;
+	Connection *connection;
+	try {
+		connection = new Connection(*wrapper->database);
+	} catch (...) {
+		return DuckDBError;
+	}
+	*out = (duckdb_connection)connection;
+	return DuckDBSuccess;
+}
+
+void duckdb_disconnect(duckdb_connection *connection) {
+	if (*connection) {
+		Connection *conn = (Connection *)*connection;
+		delete conn;
+		*connection = nullptr;
+	}
+}
+
+duckdb_state duckdb_query(duckdb_connection connection, const char *query, duckdb_result *out) {
+	Connection *conn = (Connection *)connection;
+	auto result = conn->Query(query);
+	return duckdb_translate_result(result.get(), out);
+}
+
+static void duckdb_destroy_column(duckdb_column column, idx_t count) {
+	if (column.data) {
+		if (column.type == DUCKDB_TYPE_VARCHAR) {
+			// varchar, delete individual strings
+			auto data = (char **)column.data;
+			for (idx_t i = 0; i < count; i++) {
+				if (data[i]) {
+					free(data[i]);
+				}
+			}
+		} else if (column.type == DUCKDB_TYPE_BLOB) {
+			// blob, delete individual blobs
+			auto data = (duckdb_blob *)column.data;
+			for (idx_t i = 0; i < count; i++) {
+				if (data[i].data) {
+					free((void *)data[i].data);
+				}
+			}
+		}
+		free(column.data);
+	}
+	if (column.nullmask) {
+		free(column.nullmask);
+	}
+	if (column.name) {
+		free(column.name);
+	}
+}
+
+void duckdb_destroy_result(duckdb_result *result) {
+	if (result->error_message) {
+		free(result->error_message);
+	}
+	if (result->columns) {
+		for (idx_t i = 0; i < result->column_count; i++) {
+			duckdb_destroy_column(result->columns[i], result->row_count);
+		}
+		free(result->columns);
+	}
+	memset(result, 0, sizeof(duckdb_result));
+}
