+++ src/main/capi/prepared-c.cpp
@@ -1,0 +1,134 @@
+#include "duckdb/main/capi_internal.hpp"
+
+using duckdb::Connection;
+using duckdb::MaterializedQueryResult;
+using duckdb::PreparedStatementWrapper;
+using duckdb::Value;
+
+duckdb_state duckdb_prepare(duckdb_connection connection, const char *query,
+                            duckdb_prepared_statement *out_prepared_statement) {
+	if (!connection || !query) {
+		return DuckDBError;
+	}
+	auto wrapper = new PreparedStatementWrapper();
+	Connection *conn = (Connection *)connection;
+	wrapper->statement = conn->Prepare(query);
+	*out_prepared_statement = (duckdb_prepared_statement)wrapper;
+	return wrapper->statement->success ? DuckDBSuccess : DuckDBError;
+}
+
+const char *duckdb_prepare_error(duckdb_prepared_statement prepared_statement) {
+	auto wrapper = (PreparedStatementWrapper *)prepared_statement;
+	if (!wrapper || !wrapper->statement || wrapper->statement->success) {
+		return nullptr;
+	}
+	return strdup(wrapper->statement->error.c_str());
+}
+
+duckdb_state duckdb_nparams(duckdb_prepared_statement prepared_statement, idx_t *nparams_out) {
+	auto wrapper = (PreparedStatementWrapper *)prepared_statement;
+	if (!wrapper || !wrapper->statement || !wrapper->statement->success) {
+		return DuckDBError;
+	}
+	*nparams_out = wrapper->statement->n_param;
+	return DuckDBSuccess;
+}
+
+static duckdb_state duckdb_bind_value(duckdb_prepared_statement prepared_statement, idx_t param_idx, Value val) {
+	auto wrapper = (PreparedStatementWrapper *)prepared_statement;
+	if (!wrapper || !wrapper->statement || !wrapper->statement->success) {
+		return DuckDBError;
+	}
+	if (param_idx > wrapper->statement->n_param) {
+		return DuckDBError;
+	}
+	if (param_idx > wrapper->values.size()) {
+		wrapper->values.resize(param_idx);
+	}
+	wrapper->values[param_idx - 1] = val;
+	return DuckDBSuccess;
+}
+
+duckdb_state duckdb_bind_boolean(duckdb_prepared_statement prepared_statement, idx_t param_idx, bool val) {
+	return duckdb_bind_value(prepared_statement, param_idx, Value::BOOLEAN(val));
+}
+
+duckdb_state duckdb_bind_int8(duckdb_prepared_statement prepared_statement, idx_t param_idx, int8_t val) {
+	return duckdb_bind_value(prepared_statement, param_idx, Value::TINYINT(val));
+}
+
+duckdb_state duckdb_bind_int16(duckdb_prepared_statement prepared_statement, idx_t param_idx, int16_t val) {
+	return duckdb_bind_value(prepared_statement, param_idx, Value::SMALLINT(val));
+}
+
+duckdb_state duckdb_bind_int32(duckdb_prepared_statement prepared_statement, idx_t param_idx, int32_t val) {
+	return duckdb_bind_value(prepared_statement, param_idx, Value::INTEGER(val));
+}
+
+duckdb_state duckdb_bind_int64(duckdb_prepared_statement prepared_statement, idx_t param_idx, int64_t val) {
+	return duckdb_bind_value(prepared_statement, param_idx, Value::BIGINT(val));
+}
+
+duckdb_state duckdb_bind_uint8(duckdb_prepared_statement prepared_statement, idx_t param_idx, uint8_t val) {
+	return duckdb_bind_value(prepared_statement, param_idx, Value::UTINYINT(val));
+}
+
+duckdb_state duckdb_bind_uint16(duckdb_prepared_statement prepared_statement, idx_t param_idx, uint16_t val) {
+	return duckdb_bind_value(prepared_statement, param_idx, Value::USMALLINT(val));
+}
+
+duckdb_state duckdb_bind_uint32(duckdb_prepared_statement prepared_statement, idx_t param_idx, uint32_t val) {
+	return duckdb_bind_value(prepared_statement, param_idx, Value::UINTEGER(val));
+}
+
+duckdb_state duckdb_bind_uint64(duckdb_prepared_statement prepared_statement, idx_t param_idx, uint64_t val) {
+	return duckdb_bind_value(prepared_statement, param_idx, Value::UBIGINT(val));
+}
+
+duckdb_state duckdb_bind_float(duckdb_prepared_statement prepared_statement, idx_t param_idx, float val) {
+	return duckdb_bind_value(prepared_statement, param_idx, Value(val));
+}
+
+duckdb_state duckdb_bind_double(duckdb_prepared_statement prepared_statement, idx_t param_idx, double val) {
+	return duckdb_bind_value(prepared_statement, param_idx, Value(val));
+}
+
+duckdb_state duckdb_bind_varchar(duckdb_prepared_statement prepared_statement, idx_t param_idx, const char *val) {
+	return duckdb_bind_value(prepared_statement, param_idx, Value(val));
+}
+
+duckdb_state duckdb_bind_varchar_length(duckdb_prepared_statement prepared_statement, idx_t param_idx, const char *val,
+                                        idx_t length) {
+	return duckdb_bind_value(prepared_statement, param_idx, Value(std::string(val, length)));
+}
+
+duckdb_state duckdb_bind_blob(duckdb_prepared_statement prepared_statement, idx_t param_idx, const void *data,
+                              idx_t length) {
+	return duckdb_bind_value(prepared_statement, param_idx, Value::BLOB((duckdb::const_data_ptr_t)data, length));
+}
+
+duckdb_state duckdb_bind_null(duckdb_prepared_statement prepared_statement, idx_t param_idx) {
+	return duckdb_bind_value(prepared_statement, param_idx, Value());
+}
+
+duckdb_state duckdb_execute_prepared(duckdb_prepared_statement prepared_statement, duckdb_result *out_result) {
+	auto wrapper = (PreparedStatementWrapper *)prepared_statement;
+	if (!wrapper || !wrapper->statement || !wrapper->statement->success) {
+		return DuckDBError;
+	}
+	auto result = wrapper->statement->Execute(wrapper->values, false);
+	D_ASSERT(result->type == QueryResultType::MATERIALIZED_RESULT);
+	auto mat_res = (MaterializedQueryResult *)result.get();
+	return duckdb_translate_result(mat_res, out_result);
+}
+
+void duckdb_destroy_prepare(duckdb_prepared_statement *prepared_statement) {
+	if (!prepared_statement) {
+		return;
+	}
+	auto wrapper = (PreparedStatementWrapper *)*prepared_statement;
+	if (wrapper) {
+		delete wrapper;
+	}
+	*prepared_statement = nullptr;
+}
